requirements.txt
kafka-python==2.1.0
pymongo==3.12.0
tweepy==3.11.0
textblob==0.15.3
flask==2.1.0
pyspark==3.2.0
